Mesleki Seçmeli 1/Bilgiye Erişim ve Arama Motorları/ödevler/2025/mulahteni/blm_3120/requirements.txt
argparse
datasets
pandas
joblib
scikit-learn
xgboost
matplotlib
seaborn
nltk
streamlit
Pillow
numpy